requirements.txt
torch==1.13.0
numpy==1.22.4
pandas==1.3.5
pyro-ppl==1.8.2
scikit-learn==1.0.2
hydra-core==1.3.1
tensorboard==2.10.1